requirements.txt
python-telegram-bot>=21.*
APScheduler==3.10.*
pytz
python-dotenv
